requirements.txt
annotated-types==0.7.0
anyio==3.7.1
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.104.1
h11==0.14.0
idna==3.10
mutagen==1.47.0
pycryptodomex==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
python-dotenv==1.0.1
python-multipart==0.0.6
requests==2.32.3
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.24.0
websockets==14.1
yt-dlp==2024.12.13
